requirements.txt
Flask==2.3.2
glom==23.3.0
waitress==2.1.2
mysql-connector-python==8.0.33
numpy==1.25.1
pandas==2.0.3
Werkzeug==2.3.6
python-dotenv==0.21.0
